+++ test/support/fruit.ex
@@ -5,7 +5,11 @@ defmodule Flop.Fruit do
   use Ecto.Schema
 
   @derive {Flop.Schema,
-           filterable: [:name, :family], sortable: [:name], default_limit: 50}
+           filterable: [:name, :family],
+           sortable: [:name],
+           default_limit: 50,
+           default_order_by: [:name],
+           default_order_directions: [:asc]}
 
   schema "fruits" do
     field :name, :string